adding simplelegend component and consolidate styles
